requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
dnspython==2.2.1
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
isort==5.10.1
mccabe==0.6.1
mongoengine==0.24.1
multidict==6.0.2
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==4.0.2
python-dotenv==0.20.0
pytz==2022.1
toml==0.10.2
yarl==1.7.2
